jang/08_django/02_drf/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
Django==3.2.13
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.0.0
idna==3.4
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
mypy==0.981
mypy-extensions==0.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.13.0
pytz==2022.4
refurb==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
